Add admin UI for price plugin configuration (#46)
* Add admin UI for price plugin configuration

- Add listPricePlugins and updatePricePlugin API functions in
  portfolio-api.ts following the identifier/description pattern
- Add price plugins admin page at /admin/plugins/price using the
  shared PluginConfigEditor component
- Add "Price plugins" card to admin dashboard with plugin name summary
- Add "Price" link to sidebar navigation under Plugins section


* Fix volume JSON decode and handle 403 as non-retryable

- Change AggBar.V from int64 to float64: the Massive API returns volume
  in scientific notation (e.g. 8.833325e+06) which cannot unmarshal to int64
- Add ErrForbidden client error for 403 responses (plan doesn't cover
  the requested timeframe)
- Map 403 to ErrNoData in the price plugin so the orchestrator skips the
  instrument instead of treating it as a transient error


---------

+++ server/plugins/massive/client/types.go
@@ -32,7 +32,7 @@ type AggBar struct {
 	H  float64 `json:"h"`  // high
 	L  float64 `json:"l"`  // low
 	C  float64 `json:"c"`  // close
-	V  int64   `json:"v"`  // volume
+	V  float64 `json:"v"`  // volume (float64: API returns scientific notation)
 	VW float64 `json:"vw"` // volume-weighted average price
 	T  int64   `json:"t"`  // Unix millisecond timestamp
 	N  int     `json:"n"`  // number of trades
